requirements.txt
boto3==1.36.21
streamlit==1.42.0
pillow==11.1.0